Service domain model adjustment
Services use a collection of strings for references to the services they
connect to. This makes the service data objects more compatible with how
the javascript needs to translate the data into graph nodes.

Services had previously used a navigation property / collection of
services to determine which services are connected. When translating
this to JSON, infinite loops are created whenever services are linked to
previous services in a circle, instead of only referencing each service
once.