Service data model altered to track parents
Parent connections are now tracked in the data model.
The parents are entries that are already tracked as connected services.
The parents are there so that they may be accessed later to determine
parent /child relationships.

The model also includes a boolean which will be used to determine
whether a selected service will trigger its parent services to be
included in the selection.