Altered Double Click
Double click is now called from WITHIN the mouse down function if the
time since the last click is short enough.  This means that the double
click does not also trigger the dragging event. The jquery version of
the dblclick event has been removed in favor of this new method.