Added Service Information partial view
The service info partial view is used to display all relevant info for a
single service. A controller method has been added to services, which
currently is populating the view with a single set of test data.

The home controller's index method points directly to the service
controller's services method. This makes sure the home page is really
the services page.