List of user Selected Services now global
To allow other js to access and use the list of user selected services,
the node selects list has been moved to global scope. IT is declared at
the entry point for our graph system, in Services.

Also, to prepare for immediate future changes, an ID property has been
added to each node. This is populated by the ID sent from the server.