Service Model changes. Added Service View Model
The service model now supports navigation properties for EF's benefit
and compatibility when generating a database. The nav properties were
previously lists, which EF cannot correctly create a db with.

A view model has been added, which is essentially the same as the
Service model BEFORE these changes. The view model will be used to
package data before JSON transmission to the client.