Node graph setup and drawing structure changes
Large changes have been made to the process in which services are sent
to the graph to be drawn and the process by which that happens.

Services are now stored in an array containing js service objects, the
model for which is analogous to the c# service data model. The model is
extended to provide client specific data retention (as with the
'selectable' property).

Also, changes to the Personal Plan service list generation have been
made to allow flexibility. The service list was populated by all
services EXCEPT the first in the list. It has been changed so that the
only services that do not show up are ones that are NOT selectable.

The service objects are now used in a new function which handles
population of the arbor system SEPARATE from the initial graph setup and
service data retrieval.

This paves the way for the ability to have only certain nodes drawn at
any given time, instead of all nodes at once.