Node graph dynamic node creation
The graph draw method at Services now will dynamically choose which
nodes and edges to add to the particle system before drawing. By
default, any node that cannot be selectable will be added to the graph.
Any connected nodes to a displayed node will also be displayed (but no
further connections are considered). Any node that is selected will also
have its connections displayed.

This is a contribution toward the goal of having nodes dynamically
'appear' and 'disappear' when users perform actions relating to the
graph.