Added more node test data
Added a few extra nodes to better test dynamic node instantiation.

A few commented out lines are added, they are there as part of a future
update. The update inexplicably broke the app so most of the work was
removed until it can be resolved, but a few lines exist (commented out)
that weren't part of the problem.