Style sheet organization
Trying to fix scaling issue with the services list inside of the
sidebar.  When zooming in and out, the list will extend past the borders
of the sidebar.  This led me to organize the styles sheet.

Fixed the issue by using 'fixed' positioning within the sidebar, however
that broke the functionality of the sliding effect.