Renderer cleanup: removed selected services array
Selected services array retained a list of the current selected
services. This was removed because the information about which services
are selected is already tracked by the services array.

The services array contains persistent data about all available
services, and each service in the array contains a bool property for
whether or not they are selected, which is why the selected services
array is redundant.