Node single click selection tweaks
A small mouse movement tolerance for single clicking a node has been
added. The new selection system has made it so that any minute movement
while clicking will result in a drag instead of a click. This new
tolerance allows a small drag to still be considered a click for
improved usability. It is difficult to determine whether the click
should be a drag or a selection, so a small, unnoticeable drag will
always happen first, though a large drag will still cause the selection
to not happen in favor of the drag event.