Cleaned up code
Took out bootstrap grid-scale system as it does not seem to be
benefiting us, which allowed me to take out bunches of unnecessary divs,
which also allowed me to clean up style sheet.  Changed the way CSS
properties are changed when service list items are selected.  Noticed a
bug when a list item is deselected, the hover color was not reset to
what it should've been.  Instead of adding or changing CSS properties, I
simply refactored the code to add or remove CSS classes.  Cleaned up
services.js.