Canvas scaling rectification
The canvas and the arbor system now scales correctly to the container
that it is in.

A special div "container" is needed to house the canvas. Through CSS it
conforms fluidly to space that it is in. The canvas and particle system
sample its size whenever the browser window size changes.

NOTE: Currently, the header and the container elements cause an extra
space at the bottom of the page and also a scroll bar. This should just
be styling issues because the container size is not being adjusted via
javascript.