Several minor changes
Moved button area on contact form to buttom of sidebar to line up better
with the button on the personal plan and so that when selecting list
items it doesn't move the button area down.

Added favicon image to title.

Moved service image to service info sidebar.  Still don't know why it's
destroying framerate.  Also, it's only 50x50 right now which might be
way to small.