Renderer SOC experiment
This is an experiment for ways to create greater SOC in the renderer
file. More work needs to be done to restore all site functionality, but
I need to devote my time elsewhere. This branch may be continued or
looked to in the future as an example of what I have already got
working/tried.