Renderer.js optimization and fix
The event that handles judging if the mouse is over a node happens on
every particle system update. The routine involved retrieving info from
a hovered node also happened every update. This caused a great
performance hit.

Now node data is only retrieved once when the node being hovered over is
NOT the same as the last node that was hovered over.