Node hover performance optimization (verify)
The hover function has been optimized slightly.
General performance has been gauged for the different operations in the
function, and the call to have the sidebar open was the slowest. This
now is only called once. More may need to be done based on tests.

See commented-out code on how performance was gauged.