Services js and renderer js architecture tweaks
Initialization for the graph was fully moved into renderer in my last
commit. Forgot to mention it. Now the list initialization for the
personal plan sidebar has been moved there and some non-dry code has
been removed as a result of this merge.

Even though the sidebar divs should ideally not be handled in renderer,
the list initialization was moved to keep everything in one logical
place for now.